+++ src/main/java/domain/PlacementValidator.java
@@ -1,5 +1,4 @@
-package domain.logic;
-import domain.models.Board;
+package domain;
 
 import java.util.List;
 
@@ -11,18 +10,14 @@ public class PlacementValidator {
     }
 
     public void validateSettlementPlacement(int nodeId) {
-        // In your design, Board has checkDistanceRule.
-        // We use that to determine if the placement is valid.
         if (!board.checkDistanceRule(nodeId)) {
             throw new IllegalPlacementException("Placement violates the distance rule.");
         }
     }
 
     public void validateInitialRoad(int edgeId, int settlementNodeId) {
-        // Fetch the two endpoints of the edge from the board
         List<Integer> endpoints = board.getRoadEndpoints(edgeId);
 
-        // If the board returns null or the list doesn't contain our node, it's illegal
         if (endpoints == null || !endpoints.contains(settlementNodeId)) {
             throw new IllegalPlacementException("Road must be connected to the newly placed settlement.");
         }
