+++ src/main/java/domain/TurnManager.java
@@ -1,4 +1,4 @@
-package domain.logic;
+package domain;
 
 public class TurnManager {
     private int currentPlayerIndex = 0;
@@ -10,17 +10,12 @@ public class TurnManager {
     }
 
     public void nextPlacement() {
-        // Increment count BEFORE logic to track how many houses were placed
         placementsCount++;
 
-        // Move to next player
-        // EXCEPT: At the end of the first round (placementsCount == numPlayers)
-        // the index stays the same for the snake draft reversal.
         if (placementsCount < numPlayers) {
             currentPlayerIndex++;
         }
         else if (placementsCount > numPlayers && placementsCount < numPlayers * 2) {
-            // After the peak of the snake, we start moving backward
             currentPlayerIndex--;
         }
     }
